+++ examples-pydantic-v1/equipment_unit_pydantic_v1/data_classes/_equipment_module.py
@@ -22,6 +22,7 @@ from ._core import (
     GraphQLCore,
     ResourcesWrite,
     T_DomainModelList,
+    as_direct_relation_reference,
     as_node_id,
     as_pygen_node_id,
     are_nodes_equal,
@@ -197,7 +198,7 @@ class EquipmentModuleWrite(DomainModelWrite):
     _view_id: ClassVar[dm.ViewId] = dm.ViewId("IntegrationTestsImmutable", "EquipmentModule", "b1cd4bf14a7a33")
 
     space: str = DEFAULT_INSTANCE_SPACE
-    node_type: Union[dm.DirectRelationReference, None] = None
+    node_type: Union[dm.DirectRelationReference, dm.NodeId, tuple[str, str], None] = None
     description: Optional[str] = None
     name: Optional[str] = None
     sensor_value: Union[TimeSeries, str, None] = None
@@ -236,7 +237,7 @@ class EquipmentModuleWrite(DomainModelWrite):
                 space=self.space,
                 external_id=self.external_id,
                 existing_version=None if allow_version_increase else self.data_record.existing_version,
-                type=self.node_type,
+                type=as_direct_relation_reference(self.node_type),
                 sources=[
                     dm.NodeOrEdgeData(
                         source=self._view_id,