+++ examples-pydantic-v1/omni_pydantic_v1/data_classes/_cdf_external_references_listed.py
@@ -21,6 +21,7 @@ from ._core import (
     GraphQLCore,
     ResourcesWrite,
     T_DomainModelList,
+    as_direct_relation_reference,
     as_node_id,
     as_pygen_node_id,
     are_nodes_equal,
@@ -192,7 +193,7 @@ class CDFExternalReferencesListedWrite(DomainModelWrite):
     _view_id: ClassVar[dm.ViewId] = dm.ViewId("pygen-models", "CDFExternalReferencesListed", "1")
 
     space: str = DEFAULT_INSTANCE_SPACE
-    node_type: Union[dm.DirectRelationReference, None] = None
+    node_type: Union[dm.DirectRelationReference, dm.NodeId, tuple[str, str], None] = None
     files: Optional[list[str]] = None
     sequences: Optional[list[str]] = None
     timeseries: Union[list[TimeSeries], list[str], None] = None
@@ -225,7 +226,7 @@ class CDFExternalReferencesListedWrite(DomainModelWrite):
                 space=self.space,
                 external_id=self.external_id,
                 existing_version=None if allow_version_increase else self.data_record.existing_version,
-                type=self.node_type,
+                type=as_direct_relation_reference(self.node_type),
                 sources=[
                     dm.NodeOrEdgeData(
                         source=self._view_id,