+++ examples-pydantic-v1/scenario_instance_pydantic_v1/client/data_classes/_scenario_instance.py
@@ -23,6 +23,7 @@ from ._core import (
     GraphQLCore,
     ResourcesWrite,
     T_DomainModelList,
+    as_direct_relation_reference,
     as_node_id,
     as_pygen_node_id,
     are_nodes_equal,
@@ -239,7 +240,7 @@ class ScenarioInstanceWrite(DomainModelWrite):
     _view_id: ClassVar[dm.ViewId] = dm.ViewId("IntegrationTestsImmutable", "ScenarioInstance", "ee2b79fd98b5bb")
 
     space: str = DEFAULT_INSTANCE_SPACE
-    node_type: Union[dm.DirectRelationReference, None] = None
+    node_type: Union[dm.DirectRelationReference, dm.NodeId, tuple[str, str], None] = None
     aggregation: Optional[str] = None
     country: Optional[str] = None
     instance: Optional[datetime.datetime] = None
@@ -294,7 +295,7 @@ class ScenarioInstanceWrite(DomainModelWrite):
                 space=self.space,
                 external_id=self.external_id,
                 existing_version=None if allow_version_increase else self.data_record.existing_version,
-                type=self.node_type,
+                type=as_direct_relation_reference(self.node_type),
                 sources=[
                     dm.NodeOrEdgeData(
                         source=self._view_id,