+++ examples-pydantic-v1/windmill_pydantic_v1/data_classes/_nacelle.py
@@ -22,6 +22,7 @@ from ._core import (
     GraphQLCore,
     ResourcesWrite,
     T_DomainModelList,
+    as_direct_relation_reference,
     as_node_id,
     as_pygen_node_id,
     are_nodes_equal,
@@ -329,7 +330,7 @@ class NacelleWrite(DomainModelWrite):
     _view_id: ClassVar[dm.ViewId] = dm.ViewId("power-models", "Nacelle", "1")
 
     space: str = DEFAULT_INSTANCE_SPACE
-    node_type: Union[dm.DirectRelationReference, None] = None
+    node_type: Union[dm.DirectRelationReference, dm.NodeId, tuple[str, str], None] = None
     acc_from_back_side_x: Union[TimeSeries, str, None] = None
     acc_from_back_side_y: Union[TimeSeries, str, None] = None
     acc_from_back_side_z: Union[TimeSeries, str, None] = None
@@ -429,7 +430,7 @@ class NacelleWrite(DomainModelWrite):
                 space=self.space,
                 external_id=self.external_id,
                 existing_version=None if allow_version_increase else self.data_record.existing_version,
-                type=self.node_type,
+                type=as_direct_relation_reference(self.node_type),
                 sources=[
                     dm.NodeOrEdgeData(
                         source=self._view_id,