+++ examples-pydantic-v1/windmill_pydantic_v1/data_classes/_power_inverter.py
@@ -21,6 +21,7 @@ from ._core import (
     GraphQLCore,
     ResourcesWrite,
     T_DomainModelList,
+    as_direct_relation_reference,
     as_node_id,
     as_pygen_node_id,
     are_nodes_equal,
@@ -186,7 +187,7 @@ class PowerInverterWrite(DomainModelWrite):
     _view_id: ClassVar[dm.ViewId] = dm.ViewId("power-models", "PowerInverter", "1")
 
     space: str = DEFAULT_INSTANCE_SPACE
-    node_type: Union[dm.DirectRelationReference, None] = None
+    node_type: Union[dm.DirectRelationReference, dm.NodeId, tuple[str, str], None] = None
     active_power_total: Union[TimeSeries, str, None] = None
     apparent_power_total: Union[TimeSeries, str, None] = None
     reactive_power_total: Union[TimeSeries, str, None] = None
@@ -229,7 +230,7 @@ class PowerInverterWrite(DomainModelWrite):
                 space=self.space,
                 external_id=self.external_id,
                 existing_version=None if allow_version_increase else self.data_record.existing_version,
-                type=self.node_type,
+                type=as_direct_relation_reference(self.node_type),
                 sources=[
                     dm.NodeOrEdgeData(
                         source=self._view_id,