+++ examples-pydantic-v1/windmill_pydantic_v1/data_classes/_sensor_position.py
@@ -21,6 +21,7 @@ from ._core import (
     GraphQLCore,
     ResourcesWrite,
     T_DomainModelList,
+    as_direct_relation_reference,
     as_node_id,
     as_pygen_node_id,
     are_nodes_equal,
@@ -272,7 +273,7 @@ class SensorPositionWrite(DomainModelWrite):
     _view_id: ClassVar[dm.ViewId] = dm.ViewId("power-models", "SensorPosition", "1")
 
     space: str = DEFAULT_INSTANCE_SPACE
-    node_type: Union[dm.DirectRelationReference, None] = None
+    node_type: Union[dm.DirectRelationReference, dm.NodeId, tuple[str, str], None] = None
     edgewise_bend_mom_crosstalk_corrected: Union[TimeSeries, str, None] = None
     edgewise_bend_mom_offset: Union[TimeSeries, str, None] = None
     edgewise_bend_mom_offset_crosstalk_corrected: Union[TimeSeries, str, None] = None
@@ -363,7 +364,7 @@ class SensorPositionWrite(DomainModelWrite):
                 space=self.space,
                 external_id=self.external_id,
                 existing_version=None if allow_version_increase else self.data_record.existing_version,
-                type=self.node_type,
+                type=as_direct_relation_reference(self.node_type),
                 sources=[
                     dm.NodeOrEdgeData(
                         source=self._view_id,